requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
dateparser==1.2.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pycryptodome==3.23.0
python-binance==1.0.30
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.10.23
requests==2.32.5
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
